moved dialog component
