+++ src/components/dialogs/BibtexExportProgressDialog.tsx
